Relocate MCP server, spend/error log, team membership, and managed file tables into the model layer
Moves the remaining standalone DB-table pydantic definitions out of
litellm/proxy/_types.py and into litellm/models/, re-exported from _types
for backwards compatibility. Also moves LiteLLM_BudgetTableFull and
LiteLLM_TeamMemberTable next to LiteLLM_BudgetTable, and completes the
models package __init__ exports.

Imports MCPAuthType/MCPTransportType in mcp_server_manager from their source
litellm.types.mcp instead of the proxy._types re-export, breaking an import
cycle that otherwise surfaced once the model layer pulled in mcp_server.
